fix(checker): split cross-file cache accounting
